moved index code into seperate app
