+++ fintrack/management/commands/sub_methods/index.py
@@ -1,6 +1,6 @@
 from django.db.backends.utils import logger
 
-from fintrack_be.functions.index.sp500 import SP500
+from index.services.index.sp500 import SP500
 
 
 def create_indices():
