+++ fintrack_be/admin/__init__.py
@@ -1,4 +1,4 @@
-from .index_admin import *
+from index.admin.index_admin import *
 
 admin.site.site_header = "FinTrack"
 admin.site.site_header = "FinTrack Admin"
