--- fintrack_be/functions/__init__.py
@@ -1,6 +1,0 @@
-from .company import *
-from .exchange import *
-from .index import *
-from .industry import *
-from .sector import *
-from .stock import *
