+++ fintrack_be/models/__init__.py
@@ -1,4 +1,1 @@
-from .index import Index
-from .index_constituents import IndexConstituents
-from .index_price import IndexPriceData
 from .position import Position
