--- fintrack_be/urls/base_urls.py
@@ -1,5 +1,0 @@
-from django.urls import path, include
-
-urlpatterns = [
-    path('index/', include('fintrack_be.urls.index_urls')),
-]
