+++ index/admin/__init__.py
@@ -1,0 +1,2 @@
+from .index_admin import *
+from .index_price_admin import *
