+++ index/admin/index_admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from fintrack_be.models import Index
+from index.models import Index
 
 
 class ConstituentInlineAdmin(admin.TabularInline):
