+++ index/csv/S&P 500-joined-closes.csv
