+++ index/models/index_constituents.py
@@ -1,6 +1,6 @@
 from django.db import models
 from stock.models import Stock
-from fintrack_be.models import Index
+from index.models import Index
 
 
 class IndexConstituents(models.Model):
@@ -11,7 +11,6 @@ class IndexConstituents(models.Model):
     class Meta:
         verbose_name = 'Index Constituents'
         verbose_name_plural = 'Indices Constituents'
-        db_table = 'index_constituents'
 
         constraints = [
             models.UniqueConstraint(fields=['constituent', 'index'], name='index_constituent')
