+++ index/serializers/__init__.py
@@ -1,0 +1,1 @@
+from .index_serializer import *
