+++ index/services/__init__.py
@@ -1,0 +1,2 @@
+from .index import *
+from .ml import *
