+++ index/services/index/index_class.py
@@ -1,6 +1,6 @@
 from django.db import IntegrityError
 
-from fintrack_be.models import Index
+from index.models import Index
 
 
 class IndexClass:
