+++ index/services/ml/featuresets.py
@@ -1,41 +1,7 @@
-import pandas as pd
 import numpy as np
 from collections import Counter
 
-from fintrack_be.models import Stock, Index, IndexConstituents
-
-
-def process_data_for_labels(ticker):
-    hm_days = 7
-
-    # Get the index that the stock has been part of the longest and get the joint
-    # csv file for that index
-    stock = Stock.objects.get(ticker=ticker)
-    stock_index = IndexConstituents.objects.filter(constituent=stock).order_by('date_joined')[0]
-    print(stock_index.index_id)
-    index = Index.objects.get(id=stock_index.index_id)
-    df = pd.read_csv('fintrack_be/csv/{}-joined-closes.csv'.format(index.name), index_col=0)
-
-    tickers = df.columns.values.tolist()
-    df.fillna(0, inplace=True)
-
-    for i in range(1, hm_days + 1):
-        df['{}_{}d'.format(ticker, i)] = (df[ticker].shift(-i) - df[ticker]) / df[ticker]
-    df.fillna(0, inplace=True)
-    print(tickers)
-    print(df)
-    return tickers, df
-
-
-def buy_sell_hold(*args):
-    cols = [c for c in args]
-    requirement = 0.02
-    for col in cols:
-        if col > requirement:
-            return 1
-        if col < -requirement:
-            return -1
-    return 0
+from index.services.ml import *
 
 
 def extract_featuresets(ticker):
